% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{png2fpng}
\alias{png2fpng}
\title{convert png to be compatible with fpng}
\usage{
png2fpng(png_path, fpng_path)
}
\arguments{
\item{png_path}{path to png file to be converted}

\item{fpng_path}{path to output png file}
}
\description{
convert png to be compatible with fpng
}
