% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readFPNG}
\alias{readFPNG}
\title{Decode from file
Reads an image from a PNG file into a raster array.}
\usage{
readFPNG(file_path)
}
\arguments{
\item{file_path}{character. file to write to}
}
\value{
array representing the FPNG image
}
\description{
Decode from file
Reads an image from a PNG file into a raster array.
}
