% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reshape_to_fpng}
\alias{reshape_to_fpng}
\title{Reshape png array
Reshape a png array that was read with \link[png:readPNG]{png::readPNG} to comply with fpng}
\usage{
reshape_to_fpng(x)
}
\arguments{
\item{x}{array representing a png image}
}
\description{
Reshape png array
Reshape a png array that was read with \link[png:readPNG]{png::readPNG} to comply with fpng
}
