% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{writeFPNG}
\alias{writeFPNG}
\title{Encode to file}
\usage{
writeFPNG(x, file_path)
}
\arguments{
\item{x}{integer matrix}

\item{file_path}{character. file to write to}
}
\value{
bool
}
\description{
Encode to file
}
